manager-service/requirements.txt
Django==4.2.7
pymysql
cryptography
djangorestframework==3.14.0
requests==2.31.0
django-cors-headers==4.3.1
